Adding code to create a password (for guests)
Inspired by information from various places, particularly XKCD.
Also XKPasswd.net
In the use of Modulo to keep the random number in a certain range, one
webpage presented the opinion that it added some bias to lower numbers,
so that should be avoided.
A webpage complained about weak Psudo-Random Number Generators, so php's
mcrypt_create_iv should be used.

I haven't decided if I want to use the _rand() I created based on those
complaints.  So it's just here for reference for now.