Fixing DataTables on registered devices page
Had to move jQuery/Bootstrap script tag from the bottom of the page to the
top, so that I don't have to include DataTables on every page.  Also
it would have been imposible init the table otherwise.
And I set the last column, where the delete button is, to not sort or
search.