mac registry - edit entries and will add fields and iot categories
Add a page for editting an entry in the mac registry.
Move registration, import, and delete actions to the new page.
Add Fields and IoT device categories which will be added to radius later.
Add a modification log to the mac registry.
Add database structure for the databases and tables.