Add facility to store a users default password.
cli script to change the password on all objects using the default password
to a new default.
add button in admin/object to show default password.
update database for default password table, and initialize for all students.
unlinked page to check if a user is still using their default password,
and prompt them to change their password if they are.