/api/set_default_password - Allow from user login and arbitrary password
Can override the default schema with an arbitrary password.
Can call with a user session rather than only from an api client session.
Return a message if not a student.
authenticate_api_client() should kill the (users) session.
Exit on errors sooner (so that there is only one error at a time).