Fix framework message system - no default for PASSWD_RESET, can't store
the message_settings table wasn't allowing null userid, which is needed for
the default setting.
The message_queue table wasn't allowing null from_uid, which can happen if no
from user is supplied.
The store_message() function didn't have the correct to_uid to be stored,
resulting in failure to store (to_uid can't be null).  The to user is now
passed to the function. (userid in the message_settings table may be null.)